Fixed documentation Moved access control enumerable into own directory
