+++ contracts/src/access/access_control/mod.rs
@@ -19,6 +19,10 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+pub mod extensions {
+    pub mod enumerable;
+}
+
 mod access_control;
 mod members;
