+++ contracts/src/access/mod.rs
@@ -21,7 +21,5 @@
 
 #[cfg(feature = "access_control")]
 pub mod access_control;
-#[cfg(feature = "access_control")]
-pub mod access_control_enumerable;
 #[cfg(feature = "ownable")]
 pub mod ownable;
