+++ contracts/src/governance/timelock_controller/mod.rs
@@ -91,7 +91,7 @@ where
     F: FnOnce(&mut T) -> Result<R, E>,
     E: From<AccessControlError>,
 {
-    if !instance.get().members.has_role(&role, &ZERO_ADDRESS.into()) {
+    if !instance.get().members.has_role(role, &ZERO_ADDRESS.into()) {
         check_role(instance, role, T::env().caller())?;
     }
     body(instance)
