+++ docs/docs/getting-started.md
@@ -9,8 +9,12 @@ Welcome to OpenBrush documentation! This documentation aims to guide you through
 
 ## What is OpenBrush
 OpenBrush is a library for smart contract development on ink!
-It provides standard contracts ([based on PSP](https://github.com/w3f/PSPs)), as well as useful contracts and macros to help you build ink! smart contracts.
+It provides standard contracts ([based on PSP](https://github.com/w3f/PSPs)), 
+as well as useful contracts and macros to help you build ink! smart contracts.
 
 ## Why OpenBrush
-OpenBrush attempts to analogize OpenZeppelin perfectly with Rust’s paradigm, enabling users to import contracts implemented by another user without problems and reuse the code.
-There was a need to have a library that can provide base implementations of ERCs and to import/reuse them by customizing their own logic.
+OpenBrush attempts to analogize OpenZeppelin perfectly with Rust’s paradigm, 
+enabling users to import contracts implemented by another user without problems 
+and reuse the code.
+There was a need to have a library that can provide base implementations of 
+ERCs and to import/reuse them by customizing their own logic.
