+++ docs/docs/smart-contracts/PSP22/Extensions/metadata.md
@@ -3,7 +3,7 @@ sidebar_position: 1
 title: PSP22 Metadata
 ---
 
-This example shows how you can reuse the implementation of [PSP22](https://github.com/Supercolony-net/openbrush-contracts/tree/main/contracts/src/token/psp22) token with the [PSP22Metadata](https://github.com/Supercolony-net/openbrush-contracts/tree/main/contracts/src/token/psp22/src/extensions/metadata.rs) extension.
+This example shows how you can reuse the implementation of [PSP22](https://github.com/Supercolony-net/openbrush-contracts/tree/main/contracts/src/token/psp22) token with the [PSP22Metadata](https://github.com/Supercolony-net/openbrush-contracts/tree/main/contracts/src/token/psp22/extensions/metadata.rs) extension.
 
 ## Step 1: Add imports and enable unstable feature
